pyproject.toml
[project]
name = "lang-graph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.24",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.8",
    "pygraphviz",
    "python-dotenv>=1.1.0",
]
